move insert to a standalone function
